Rename customViewsHeader file to issueViewsHeader
